extract UserPreferencesManager from AGUser
