+++ platform-core-module/src/main/java/com/softicar/platform/core/module/user/preferences/UserSpecificDomPreferences.java
@@ -2,8 +2,6 @@ package com.softicar.platform.core.module.user.preferences;
 
 import com.softicar.platform.core.module.user.AGUser;
 import com.softicar.platform.core.module.user.CurrentUser;
-import com.softicar.platform.core.module.user.UserPreferences;
-import com.softicar.platform.core.module.user.UserPreferencesPreferredPopupPlacement;
 import com.softicar.platform.dom.elements.popup.placement.strategy.IDomPopupPlacementStrategy;
 import com.softicar.platform.dom.user.DomDefaultPreferences;
 import com.softicar.platform.dom.user.IDomPreferences;
