+++ platform-core-module/src/main/java/com/softicar/platform/core/module/user/profile/UserProfilePreferencesDiv.java
@@ -5,8 +5,8 @@ import com.softicar.platform.common.core.interfaces.IRefreshable;
 import com.softicar.platform.core.module.CoreI18n;
 import com.softicar.platform.core.module.CoreTestMarker;
 import com.softicar.platform.core.module.user.CurrentUser;
-import com.softicar.platform.core.module.user.UserPreferences;
-import com.softicar.platform.core.module.user.UserPreferencesPreferredPopupPlacement;
+import com.softicar.platform.core.module.user.preferences.UserPreferences;
+import com.softicar.platform.core.module.user.preferences.UserPreferencesPreferredPopupPlacement;
 import com.softicar.platform.dom.elements.DomDiv;
 import com.softicar.platform.dom.elements.bar.DomActionBar;
 import com.softicar.platform.dom.elements.button.DomButton;
